Win32 & UNIX integration
git-svn-id: svn://svn.zabbix.com/branches/zabbix-1_2_dev@2970 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ src/zabbix_agent/zbxconf.h
@@ -27,8 +27,8 @@ extern char	*CONFIG_STAT_FILE_TMP;
 /* extern int	CONFIG_NOTIMEWAIT;		*/
 extern int	CONFIG_DISABLE_ACTIVE;
 extern int	CONFIG_ENABLE_REMOTE_COMMANDS;
-extern unsigned short	CONFIG_LISTEN_PORT;
-extern unsigned short	CONFIG_SERVER_PORT;
+extern int	CONFIG_LISTEN_PORT;
+extern int	CONFIG_SERVER_PORT;
 extern int	CONFIG_REFRESH_ACTIVE_CHECKS;
 extern char	*CONFIG_LISTEN_IP;
 extern int	CONFIG_LOG_LEVEL;
